image_video_mcp/pyproject.toml
[project]
name = "image-video-mcp"
version = "0.1.0"
description = "基于 FastMCP 的图像和视频生成 MCP 服务"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.13.1",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "google-genai>=0.2.0",
    "pillow>=10.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/image_video_mcp"]

[project.scripts]
image-video-mcp = "image_video_mcp.main:main"
